requirements.txt
streamlit
openai
python-dotenv
trimesh
plotly 